GH-457 Move ErrorResponse and PermissionNotFoundException to shared project
