--- region-connectors/region-connector-aiida/src/main/java/energy/eddie/regionconnector/aiida/dtos/ErrorResponse.java
@@ -1,12 +1,0 @@
-package energy.eddie.regionconnector.aiida.dtos;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
-
-import java.util.List;
-
-public record ErrorResponse(
-        @NotNull
-        @JsonProperty List<String> errors
-) {
-}
