--- region-connectors/region-connector-dk-energinet/src/main/java/energy/eddie/regionconnector/dk/energinet/dtos/ErrorResponse.java
@@ -1,12 +1,0 @@
-package energy.eddie.regionconnector.dk.energinet.dtos;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
-public record ErrorResponse(
-        @NotNull
-        @JsonProperty List<String> errors
-) {
-}
