+++ region-connectors/region-connector-es-datadis/src/main/java/energy/eddie/regionconnector/es/datadis/web/PermissionController.java
@@ -3,8 +3,8 @@ package energy.eddie.regionconnector.es.datadis.web;
 import energy.eddie.api.v0.ConnectionStatusMessage;
 import energy.eddie.api.v0.process.model.StateTransitionException;
 import energy.eddie.regionconnector.es.datadis.dtos.PermissionRequestForCreation;
-import energy.eddie.regionconnector.es.datadis.dtos.exceptions.PermissionNotFoundException;
 import energy.eddie.regionconnector.es.datadis.services.PermissionRequestService;
+import energy.eddie.regionconnector.shared.exceptions.PermissionNotFoundException;
 import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
