+++ region-connectors/shared/src/main/java/energy/eddie/regionconnector/shared/dtos/ErrorResponse.java
@@ -1,12 +1,10 @@
-package energy.eddie.regionconnector.es.datadis.dtos;
+package energy.eddie.regionconnector.shared.dtos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
 
 public record ErrorResponse(
-        @NotNull
         @JsonProperty List<String> errors
 ) {
 }
